% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexity_index.R
\name{calculate.complexity.index}
\alias{calculate.complexity.index}
\title{calculate.complexity.index}
\usage{
calculate.complexity.index(spectra)
}
\arguments{
\item{spectra}{Spectral signatures of fluorophores, normalized between 0 and 1,
with fluorophores in rows and detectors (channels) in columns. More than one
spectra are needed to calculate anything here.}
}
\value{
The complexity index (condition number) for the
spectral flow cytometry panel.
}
\description{
Calculates the complexity index for a set of fluorophores (a spectral panel).
}
