% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_similarity_matrix.R
\name{calculate.similarity.matrix}
\alias{calculate.similarity.matrix}
\title{calculate.similarity.matrix}
\usage{
calculate.similarity.matrix(spectra)
}
\arguments{
\item{spectra}{Spectral signatures of fluorophores, normalized between 0 and 1,
with fluorophores in rows and detectors (channels) in columns. More than one
spectra are needed to calculate anything here.}
}
\value{
The similarity matrix for the spectral flow cytometry panel.
}
\description{
Calculates the similarity matrix (set of cosine similarity values)
for a set of fluorophores (a spectral panel).
}
