% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ssm.R
\name{calculate.ssm}
\alias{calculate.ssm}
\title{calculate.ssm}
\usage{
calculate.ssm(spectra)
}
\arguments{
\item{spectra}{Spectral signatures of fluorophores, normalized between 0 and 1,
with fluorophores in rows and detectors (channels) in columns. More than one
spectra are needed to calculate anything here.}
}
\value{
The spillover spread matrix (ssm) for the
spectral flow cytometry panel. Spread from row fluor into column fluor.
}
\description{
Calculates a spillover spread matrix (ssm) for a set of fluorophores.
}
