% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_matrix_heatmap.R
\name{similarity.matrix.heatmap}
\alias{similarity.matrix.heatmap}
\title{Similarity Matrix Heatmap}
\usage{
similarity.matrix.heatmap(
  spectra,
  file.name = "Similarity Matrix.jpg",
  figure.width = 8,
  figure.height = 6
)
}
\arguments{
\item{spectra}{Data frame containing spectral data. Fluorophores in rows,
detectors in columns.}

\item{file.name}{Character string. Default is "Similarity Matrix.jpg". Provide
a file path to save somewhere other than the working directory.}

\item{figure.width}{Numeric. Default is 8}

\item{figure.height}{Numeric. Default is 6}
}
\value{
Saves the heatmap plot as a JPEG file in the working directory or in
the path provided as an argument to file.name.
}
\description{
This function plots the similarity matrix as a heatmap and saves
it as a JPEG file. It also calculates and displays the complexity index.
}
