% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_heatmap.R
\name{spectral.heatmap}
\alias{spectral.heatmap}
\title{Spectral Heatmap}
\usage{
spectral.heatmap(
  spectra,
  file.name = "Spectral Heatmap.jpg",
  legend.label = "Intensity"
)
}
\arguments{
\item{spectra}{Matrix or dataframe containing spectral data (fluorophores x detectors).}

\item{file.name}{Character. Output file name. Default is "Spectral Heatmap.jpg".
Path may be supplied.}

\item{legend.label}{Character string that will appear on the heatmap legend.}
}
\value{
Saves the heatmap plot as a JPEG file in the specified directory.
}
\description{
This function plots a spectral matrix as a heatmap and saves it
as a JPEG file.
}
